requirements.txt
GitPython==3.1.37
minio==7.1.17
fire==0.5.0
requests==2.31.0
junitparser==3.1.0
